♻️ Restructure Styling component files
