+++ modules/ppcp-settings/resources/js/Components/Screens/Settings/Components/Styling/Layout/StylingSectionWithRadiobuttons.js
@@ -1,7 +1,7 @@
 import { RadioControl } from '@wordpress/components';
 import classNames from 'classnames';
 
-import HStack from '../../../../ReusableComponents/HStack';
+import HStack from '../../../../../ReusableComponents/HStack';
 import StylingSection from './StylingSection';
 
 const StylingSectionWithRadiobuttons = ( {
